requirements.txt
python-pptx
pdfplumber
google-generativeai
python-dotenv
streamlit

